pyproject.toml
[project]
name = "impulse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "debugpy>=1.8.14",
    "maya-stubs>=0.4.1",
    "ruff>=0.9.6",
]
[tool.ruff]
line-length = 100

[tool.pyright]
extraPaths = ["/usr/autodesk/ApplicationPlugins/ngskintools2/Contents/scripts"]
#stubPath = "./.stubs/"
typeCheckingMode = "off"